Add basic functionality and documentation
Updated README.md to include a description of what the program does and why it may be useful.  Also included example_data.csv and example.JPG to improve the likelihood that someone will understand what the tool does and how to use it.

Updated .gitignore to include .spec files produced by PyInstaller

The wxPython code in gui.py could probably be a lot better than it is but I'm a new wxPython user and I really want to get this tool out while the October 2017 Starbone Alpha round is still running.  I'm hoping to get some feedback or grief about the program's utility.